requirements.txt
# python=3.12
python-dotenv==1.1.0
requests==2.32.3
discord.py[voice]==2.5.2
pyyaml==6.0.2
openai==1.83.0
yt-dlp==2025.5.22
beautifulsoup4==4.13.4
selenium==4.33.0
webdriver-manager==4.0.2
